BACKLOG-16831: Fixed confirmation dialog, removed formikRef
