+++ src/javascript/Api/ContentEditorModal.jsx
@@ -3,7 +3,6 @@ import {Constants} from '~/ContentEditor.constants';
 import {ContentEditor} from '~/ContentEditor';
 import {Dialog, IconButton} from '@material-ui/core';
 import styles from './ContentEditorApi.scss';
-import {FormikProvider} from 'formik';
 import EditPanelDialogConfirmation from '~/EditPanel/EditPanelDialogConfirmation/EditPanelDialogConfirmation';
 import Slide from '@material-ui/core/Slide';
 import PropTypes from 'prop-types';
@@ -32,8 +31,6 @@ export const ContentEditorModal = ({editorConfig, setEditorConfig}) => {
     const notificationContext = useNotifications();
     const [confirmationConfig, setConfirmationConfig] = useState(false);
 
-    // Todo get rid of ref in envProps, move dialog in subtree
-    const formikRef = useRef();
     const dirtyRef = useRef(false);
     const needRefresh = useRef(false);
     const dispatch = useDispatch();
@@ -49,7 +46,6 @@ export const ContentEditorModal = ({editorConfig, setEditorConfig}) => {
 
     // Standalone env props
     const envProps = {
-        formikRef,
         dirtyRef,
         back: () => {
             setEditorConfig(false);
@@ -134,7 +130,14 @@ export const ContentEditorModal = ({editorConfig, setEditorConfig}) => {
             envProps.back();
         },
         isModal: true,
-        isFullscreen: editorConfig?.isFullscreen
+        isFullscreen: editorConfig?.isFullscreen,
+        confirmationDialog: confirmationConfig && (
+            <EditPanelDialogConfirmation
+                isOpen
+                actionCallback={() => setEditorConfig(false)}
+                onCloseDialog={() => setConfirmationConfig(false)}
+            />
+        )
     };
 
     const classes = editorConfig.isFullscreen ? {
@@ -155,15 +158,6 @@ export const ContentEditorModal = ({editorConfig, setEditorConfig}) => {
                 onClose={() => (envProps.dirtyRef.current) ? setConfirmationConfig(true) : setEditorConfig(false)}
                 onRendered={() => window.focus()}
         >
-            {confirmationConfig && (
-                <FormikProvider value={formikRef.current}>
-                    <EditPanelDialogConfirmation
-                        isOpen={open}
-                        actionCallback={envProps.back}
-                        onCloseDialog={() => setConfirmationConfig(false)}
-                    />
-                </FormikProvider>
-            )}
             <ContentEditor mode={editorConfig.mode}
                            uuid={editorConfig.uuid}
                            lang={editorConfig.lang}
