+++ src/javascript/Create/CreateForm/create.action.js
@@ -5,7 +5,7 @@ import {ComponentRendererContext} from '@jahia/ui-extender';
 import * as PropTypes from 'prop-types';
 import {useFormikContext} from 'formik';
 import {useContentEditorConfigContext, useContentEditorContext} from '~/ContentEditor.context';
-import {useKeydownListener} from '~/utils/getKeydownListener';
+import {useKeydownListener} from '~/utils/useKeydownListener';
 import {adaptCreateFormData} from '~/Create/Create.adapter';
 import {useTranslation} from 'react-i18next';
 import {useContentEditorSectionContext} from '~/ContentEditorSection/ContentEditorSection.context';
@@ -20,7 +20,7 @@ const Create = ({createAnother, render: Render, loading: Loading, ...otherProps}
     const [clicked, setClicked] = useState(false);
     const {sections} = useContentEditorSectionContext();
 
-    useKeydownListener((event, formik) => {
+    useKeydownListener(event => {
         if (mode !== Constants.routes.baseCreateRoute) {
             return;
         }
