+++ src/javascript/utils/useKeydownListener.js
@@ -1,0 +1,10 @@
+import {useEffect} from 'react';
+
+export const useKeydownListener = handler => {
+    useEffect(() => {
+        addEventListener('keydown', handler);
+        return () => {
+            removeEventListener('keydown', handler);
+        };
+    }, [handler]);
+};
